backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4
yahooquery==2.3.7
ta==0.11.0
scikit-learn==1.5.0
statsmodels==0.14.2
pyarrow==16.0.0
aiofiles==23.2.1
httpx==0.27.0
websockets==12.0
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
cryptography==42.0.8
alpaca-py==0.28.0
structlog==24.1.0
pyyaml==6.0.1
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-cov==5.0.0
anyio==4.3.0
aiosqlite==0.20.0
greenlet==3.0.3
